refactor(webui): split state machine helpers (transitions, theme-identity)
useRendererHostMachine.ts(325줄)를 use-host-machine/ 폴더 3 파일로
분해한다.

- transitions.ts: HostStatus 타입, FADE_OUT/THEME_TRANSITION/FADE_IN_MS
  상수, errorMessage helper, classForStatus(status -> CSS class) 매트릭스.
- theme-identity.ts: evaluateTheme, themeIdentity, sortedTokens.
  validateTheme를 한 번 감싸 try/catch + console.warn 처리.
- use-host-machine.ts: 메인 hook 본체.
- index.ts: re-export.

RenderedView.tsx의 import 경로를 ./use-host-machine/index.js로 갱신.
또한 F1에서 rendererRuntime.tsx 삭제 시 끊겼던 RendererLayerId
import도 @/client/entities/renderer/bundle/index.js로 정정한다.

이름 변경(useRendererHostMachine -> useRendererSurfaceMachine 등)은 G1.

issue #144 Phase F5.
